refactor(library): Port-Gruppen-Helfer aus LibraryPanel auslagern (#468)
LibraryPanel.tsx 1454 → 1430 Zeilen. Reine Port-Gruppen-Helfer
(PortGroupDraft-Typ, defaultGroup, buildPorts — kein React) → neue
libraryPanelHelpers.ts. Verhaltensneutraler Code-Move.

tsc 0, eslint 0, build 0; Headless-Boot: Library-Panel rendert,
0 Console-Errors.

Refs #468
