+++ src/renderer/components/Library/libraryPanelHelpers.ts
@@ -1,0 +1,31 @@
+// #468 — Port-Gruppen-Helfer aus LibraryPanel ausgelagert (rein, kein React).
+import { v4 as uuidv4 } from 'uuid'
+import type { ConnectorType, Port } from '../../types/equipment'
+
+export interface PortGroupDraft {
+  id: string
+  direction: 'in' | 'out'
+  count: number
+  connectorType: ConnectorType
+  label: string
+}
+
+export const defaultGroup = (direction: 'in' | 'out'): PortGroupDraft => ({
+  id: uuidv4(),
+  direction,
+  count: 1,
+  connectorType: 'Custom',
+  label: direction === 'in' ? 'Input' : 'Output',
+})
+
+export const buildPorts = (groups: PortGroupDraft[], direction: 'in' | 'out'): Port[] => {
+  const filtered = groups.filter((group) => group.direction === direction)
+  return filtered.flatMap((group) =>
+    Array.from({ length: Math.max(0, group.count) }, (_item, index) => ({
+      id: uuidv4(),
+      name: `${group.label} ${index + 1}`,
+      type: group.connectorType,
+      connectorType: group.connectorType,
+    })),
+  )
+}
